requirements.txt
et-xmlfile==1.1.0
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.24.3
openpyxl==3.1.2
openpyxl-stubs==0.1.25
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
xlrd==2.0.1
